requirements.txt
# ============================================================
# IndiaLexABSA — Requirements
# Python 3.10+
# ============================================================

# ── Core ML / NLP ────────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.38.0
datasets>=2.18.0
sentence-transformers>=2.6.0
tokenizers>=0.15.0
accelerate>=0.27.0
peft>=0.9.0
evaluate>=0.4.1
seqeval>=1.2.2

# ── Explainability ────────────────────────────────────────────
shap>=0.44.0
captum>=0.7.0

# ── Hyperparameter Optimization ───────────────────────────────
optuna>=3.5.0
optuna-integration>=3.5.0

# ── Experiment Tracking ───────────────────────────────────────
wandb>=0.16.0

# ── PDF Processing ────────────────────────────────────────────
pymupdf>=1.23.0
pdfplumber>=0.10.0
pypdf>=4.0.0

# ── NLP Utilities ────────────────────────────────────────────
spacy>=3.7.0
nltk>=3.8.1
langdetect>=1.0.9
wordcloud>=1.9.3
scikit-learn>=1.4.0
scipy>=1.12.0
numpy>=1.26.0
pandas>=2.2.0

# ── Vector Database ──────────────────────────────────────────
chromadb>=0.4.22

# ── Graph Processing ─────────────────────────────────────────
networkx>=3.2.0

# ── OpenAI ───────────────────────────────────────────────────
openai>=1.14.0
tiktoken>=0.6.0

# ── HuggingFace Hub ──────────────────────────────────────────
huggingface-hub>=0.21.0

# ── Web Scraping ─────────────────────────────────────────────
selenium>=4.18.0
webdriver-manager>=4.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
aiohttp>=3.9.0

# ── Dashboard ────────────────────────────────────────────────
streamlit>=1.32.0
streamlit-extras>=0.4.0
plotly>=5.20.0
matplotlib>=3.8.0
Pillow>=10.2.0

# ── PDF Report Generation ────────────────────────────────────
reportlab>=4.1.0
svglib>=1.5.1

# ── Database / Storage ───────────────────────────────────────
sqlalchemy>=2.0.0

# ── Config Management ────────────────────────────────────────
python-dotenv>=1.0.0
pyyaml>=6.0.1
omegaconf>=2.3.0

# ── Testing ──────────────────────────────────────────────────
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# ── Utilities ────────────────────────────────────────────────
tqdm>=4.66.0
rich>=13.7.0
loguru>=0.7.0
tenacity>=8.2.0
joblib>=1.3.0
python-dateutil>=2.8.2
